+++ src/main/java/com/ia/transaction/loader/DesjardinsCCLoader.java
@@ -2,13 +2,11 @@ package com.ia.transaction.loader;
 
 import com.ia.transaction.mapper.TransactionMapper;
 import com.ia.transaction.model.DesjardinsCCTransaction;
-import com.ia.transaction.parser.TransactionParser;
 import com.ia.transaction.repository.TransactionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 import java.io.File;
-import java.util.List;
 
 /**
  * Desjardins credit card implementation of the TransactionLoader interface
@@ -21,7 +19,7 @@ public class DesjardinsCCLoader implements  TransactionLoader<File, DesjardinsCC
     private final TransactionMapper<File, DesjardinsCCTransaction> mapper;
 
     @Override
-    public int load(File source) {
+    public long load(File source) {
         return load(source, repository, mapper);
     }
 }
